Track gamepad changes in DisplayView
