+++ src/screens/home/displays/helpers/InputState.ts
@@ -7,6 +7,7 @@ import {
 } from 'nighthouse/browser';
 
 export interface InputState {
+  gamepadCount?: number;
   lastMouseEvent?: MouseEvent;
   lastKeyEvent?: KeyEvent | LegacyKeyEvent;
   lastControllerEvent?: GamepadEvent | LegacyControllerEvent;
